compilers/cc: long -> int32 for ieeedtod

+++ compilers/cc/pswt.c
@@ -193,5 +193,5 @@ ieeedtod(Ieee *ieee, double native)
 	fr = modf(fr*f, &ho);
 	ieee->l = ho;
 	ieee->l <<= 16;
-	ieee->l |= (long)(fr*f);
+	ieee->l |= (int32)(fr*f);
 }
